+++ src/wallet/puzzles/genesis-by-coin-id-with-0.clvm.hex
