+++ src/wallet/puzzles/test_cc.py
@@ -54,11 +54,7 @@ def int_as_bytes32(v: int) -> bytes32:
     return v.to_bytes(32, byteorder="big")
 
 
-def generate_farmed_coin(
-    block_index: int,
-    puzzle_hash: bytes32,
-    amount: int,
-) -> Coin:
+def generate_farmed_coin(block_index: int, puzzle_hash: bytes32, amount: int,) -> Coin:
     """
     Generate a (fake) coin which can be used as a starting point for a chain
     of coin tests.
@@ -166,10 +162,7 @@ def test_spend_through_n(mod_code, coin_checker_for_farmed_coin, n):
     inner_puzzle_solution = Program.to(output_conditions)
 
     spend_bundle = spend_bundle_for_spendable_ccs(
-        mod_code,
-        genesis_coin_checker,
-        spendable_cc_list,
-        [inner_puzzle_solution],
+        mod_code, genesis_coin_checker, spendable_cc_list, [inner_puzzle_solution],
     )
 
     debug_spend_bundle(spend_bundle)
@@ -194,10 +187,7 @@ def test_spend_through_n(mod_code, coin_checker_for_farmed_coin, n):
     ]
 
     spend_bundle = spend_bundle_for_spendable_ccs(
-        mod_code,
-        genesis_coin_checker,
-        spendable_cc_list,
-        inner_solutions,
+        mod_code, genesis_coin_checker, spendable_cc_list, inner_solutions,
     )
 
     debug_spend_bundle(spend_bundle)
